+++ inc/tools.php
@@ -128,7 +128,7 @@ add_filter( 'cptui_get_tabs', 'cptui_tools_tabs', 10, 2 );
  */
 function cptui_tools() {
 
-	$tab = '';
+	$tab = 'post_types';
 	if ( ! empty( $_GET ) ) {
 		if ( ! empty( $_GET['action'] ) && 'taxonomies' === $_GET['action'] ) {
 			$tab = 'taxonomies';
@@ -136,8 +136,6 @@ function cptui_tools() {
 			$tab = 'get_code';
 		} elseif ( ! empty( $_GET['action'] ) && 'debuginfo' === $_GET['action'] ) {
 			$tab = 'debuginfo';
-		} else {
-			$tab = 'post_types';
 		}
 	}
 
@@ -266,6 +264,7 @@ function cptui_get_single_taxonomy_registery( $taxonomy = [] ) {
 	$public             = isset( $taxonomy['public'] ) ? disp_boolean( $taxonomy['public'] ) : 'true';
 	$publicly_queryable = isset( $taxonomy['publicly_queryable'] ) ? disp_boolean( $taxonomy['publicly_queryable'] ) : disp_boolean( $taxonomy['public'] );
 	$show_in_quick_edit = isset( $taxonomy['show_in_quick_edit'] ) ? disp_boolean( $taxonomy['show_in_quick_edit'] ) : disp_boolean( $taxonomy['show_ui'] );
+	$show_tagcloud      = isset( $taxonomy['show_tagcloud'] ) ? disp_boolean( $taxonomy['show_tagcloud'] ) : disp_boolean( $taxonomy['show_ui'] );
 
 	$show_in_menu = ( ! empty( $taxonomy['show_in_menu'] ) && false !== get_disp_boolean( $taxonomy['show_in_menu'] ) ) ? 'true' : 'false';
 	if ( empty( $taxonomy['show_in_menu'] ) ) {
@@ -343,6 +342,7 @@ foreach ( $taxonomy['labels'] as $key => $label ) {
 		"rewrite" => <?php echo $rewrite; ?>,
 		"show_admin_column" => <?php echo $taxonomy['show_admin_column']; ?>,
 		"show_in_rest" => <?php echo $show_in_rest; ?>,
+		"show_tagcloud" => <?php echo $show_tagcloud; ?>,
 		"rest_base" => "<?php echo $rest_base; ?>",
 		"rest_controller_class" => "<?php echo $rest_controller_class; ?>",
 		"show_in_quick_edit" => <?php echo $show_in_quick_edit; ?>,
@@ -617,7 +617,7 @@ function cptui_get_single_post_type_registery( $post_type = [] ) {
  * @return mixed false on nothing to do, otherwise void.
  */
 function cptui_import_types_taxes_settings( $postdata = [] ) {
-	if ( ! isset( $postdata['cptui_post_import'] ) && ! isset( $postdata['cptui_tax_import'] ) ) {
+	if ( ! isset( $postdata['cptui_post_import'] ) && ! isset( $postdata['cptui_tax_import'] ) && ! array_key_exists( 'delete', $postdata ) ) {
 		return false;
 	}
 
@@ -659,13 +659,15 @@ function cptui_import_types_taxes_settings( $postdata = [] ) {
 		$postdata['cptui_tax_import'] = $third_party_taxonomy_data;
 	}
 
-	if ( ! empty( $postdata['cptui_post_import'] ) ) {
-		$cpt_data = stripslashes_deep( trim( $postdata['cptui_post_import'] ) );
-		$settings = json_decode( $cpt_data, true );
+	if ( ! empty( $postdata['cptui_post_import'] ) || ( isset( $postdata['delete'] ) && 'type_true' === $postdata['delete'] ) ) {
+		$settings = null;
+		if ( ! empty( $postdata['cptui_post_import'] ) ) {
+			$settings = $postdata['cptui_post_import'];
+		}
 
 		// Add support to delete settings outright, without accessing database.
 		// Doing double check to protect.
-		if ( null === $settings && '{""}' === $cpt_data ) {
+		if ( null === $settings && ( isset( $postdata['delete'] ) && 'type_true' === $postdata['delete'] ) ) {
 
 			/**
 			 * Filters whether or not 3rd party options were deleted successfully within post type import.
@@ -706,13 +708,15 @@ function cptui_import_types_taxes_settings( $postdata = [] ) {
 		if ( $success ) {
 			$status = 'import_success';
 		}
-	} elseif ( ! empty( $postdata['cptui_tax_import'] ) ) {
-		$tax_data = stripslashes_deep( trim( $postdata['cptui_tax_import'] ) );
-		$settings = json_decode( $tax_data, true );
+	} elseif ( ! empty( $postdata['cptui_tax_import'] ) || ( isset( $postdata['delete'] ) && 'tax_true' === $postdata['delete'] ) ) {
+		$settings = null;
 
+		if ( ! empty( $postdata['cptui_tax_import'] ) ) {
+			$settings = $postdata['cptui_tax_import'];
+		}
 		// Add support to delete settings outright, without accessing database.
 		// Doing double check to protect.
-		if ( null === $settings && '{""}' === $tax_data ) {
+		if ( null === $settings && ( isset( $postdata['delete'] ) && 'tax_true' === $postdata['delete'] ) ) {
 
 			/**
 			 * Filters whether or not 3rd party options were deleted successfully within taxonomy import.
@@ -974,18 +978,16 @@ function cptui_render_debuginfo_section() {
  * @param string $tab Current tab to display.
  */
 function cptui_render_tools( $tab ) {
-	if ( isset( $tab ) ) {
-		if ( 'post_types' === $tab || 'taxonomies' === $tab ) {
-			cptui_render_posttypes_taxonomies_section();
-		}
+	if ( 'post_types' === $tab || 'taxonomies' === $tab ) {
+		cptui_render_posttypes_taxonomies_section();
+	}
 
-		if ( 'get_code' === $tab ) {
-			cptui_render_getcode_section();
-		}
+	if ( 'get_code' === $tab ) {
+		cptui_render_getcode_section();
+	}
 
-		if ( 'debuginfo' === $tab ) {
-			cptui_render_debuginfo_section();
-		}
+	if ( 'debuginfo' === $tab ) {
+		cptui_render_debuginfo_section();
 	}
 }
 add_action( 'cptui_tools_sections', 'cptui_render_tools' );
@@ -1001,7 +1003,42 @@ function cptui_do_import_types_taxes() {
 	     ( ! empty( $_POST['cptui_post_import'] ) && isset( $_POST['cptui_post_import'] ) ) ||
 	     ( ! empty( $_POST['cptui_tax_import'] ) && isset( $_POST['cptui_tax_import'] ) )
 	) {
-		$success = cptui_import_types_taxes_settings( $_POST );
+		$data              = [];
+		$decoded_post_data = null;
+		$decoded_tax_data  = null;
+		if ( ! empty( $_POST['cptui_post_import'] ) ) {
+			$decoded_post_data = json_decode( stripslashes_deep( trim( $_POST['cptui_post_import'] ) ), true );
+		}
+
+		if ( ! empty( $_POST['cptui_tax_import'] ) ) {
+			$decoded_tax_data = json_decode( stripslashes_deep( trim( $_POST['cptui_tax_import'] ) ), true );
+		}
+
+		if (
+			empty( $decoded_post_data ) &&
+			empty( $decoded_tax_data ) &&
+			(
+				! empty( $_POST['cptui_post_import'] ) && '{""}' !== stripslashes_deep( trim( $_POST['cptui_post_import'] ) )
+			) &&
+			(
+				! empty( $_POST['cptui_tax_import'] ) && '{""}' !== stripslashes_deep( trim( $_POST['cptui_tax_import'] ) )
+			)
+		) {
+			return;
+		}
+		if ( null !== $decoded_post_data ) {
+			$data['cptui_post_import'] = $decoded_post_data;
+		}
+		if ( null !== $decoded_tax_data ) {
+			$data['cptui_tax_import'] = $decoded_tax_data;
+		}
+		if ( ! empty( $_POST['cptui_post_import'] ) && '{""}' === stripslashes_deep( trim( $_POST['cptui_post_import'] ) ) ) {
+			$data['delete'] = 'type_true';
+		}
+		if ( ! empty( $_POST['cptui_tax_import'] ) && '{""}' === stripslashes_deep( trim( $_POST['cptui_tax_import'] ) ) ) {
+			$data['delete'] = 'tax_true';
+		}
+		$success = cptui_import_types_taxes_settings( $data );
 		add_action( 'admin_notices', "cptui_{$success}_admin_notice" );
 	}
 }