+++ inc/utility.php
@@ -108,14 +108,14 @@ function cptui_footer( $original = '' ) {
 		__( '%s version %s by %s', 'custom-post-type-ui' ),
 		__( 'Custom Post Type UI', 'custom-post-type-ui' ),
 		CPTUI_VERSION,
-		'<a href="https://webdevstudios.com" target="_blank">WebDevStudios</a>'
+		'<a href="https://webdevstudios.com" target="_blank" rel="noopener">WebDevStudios</a>'
 	) . ' - ' .
 	sprintf(
-		'<a href="http://wordpress.org/support/plugin/custom-post-type-ui" target="_blank">%s</a>',
+		'<a href="http://wordpress.org/support/plugin/custom-post-type-ui" target="_blank" rel="noopener">%s</a>',
 		__( 'Support forums', 'custom-post-type-ui' )
 	) . ' - ' .
 	sprintf(
-		'<a href="https://wordpress.org/plugins/custom-post-type-ui/reviews/" target="_blank">%s</a>',
+		'<a href="https://wordpress.org/plugins/custom-post-type-ui/reviews/" target="_blank" rel="noopener">%s</a>',
 		sprintf(
 			// translators: Placeholder will hold `<abbr>` tag for CPTUI.
 			__( 'Review %s', 'custom-post-type-ui' ),
@@ -129,7 +129,7 @@ function cptui_footer( $original = '' ) {
 	__( 'Follow on Twitter:', 'custom-post-type-ui' ) .
 	sprintf(
 		' %s',
-		'<a href="https://twitter.com/webdevstudios" target="_blank">WebDevStudios</a>'
+		'<a href="https://twitter.com/webdevstudios" target="_blank" rel="noopener">WebDevStudios</a>'
 	);
 }
 add_filter( 'admin_footer_text', 'cptui_footer' );
@@ -571,11 +571,17 @@ function cptui_admin_notices_helper( $message = '', $success = true ) {
  */
 function cptui_get_object_from_post_global() {
 	if ( isset( $_POST['cpt_custom_post_type']['name'] ) ) {
-		return sanitize_text_field( $_POST['cpt_custom_post_type']['name'] );
+		$type_item = filter_input( INPUT_POST, 'cpt_custom_post_type', FILTER_SANITIZE_STRING, FILTER_REQUIRE_ARRAY );
+		if ( $type_item ) {
+			return sanitize_text_field( $type_item['name'] );
+		}
 	}
 
 	if ( isset( $_POST['cpt_custom_tax']['name'] ) ) {
-		return sanitize_text_field( $_POST['cpt_custom_tax']['name'] );
+		$tax_item = filter_input( INPUT_POST, 'cpt_custom_tax', FILTER_SANITIZE_STRING, FILTER_REQUIRE_ARRAY );
+		if ( $tax_item ) {
+			return sanitize_text_field( $tax_item['name'] );
+		}
 	}
 
 	return esc_html__( 'Object', 'custom-post-type-ui' );
@@ -872,7 +878,7 @@ function cptui_get_cptui_post_type_object( $post_type = '' ) {
 	if ( array_key_exists( $post_type, $post_types ) ) {
 		return $post_types[ $post_type ];
 	}
-	return '';
+	return [];
 }
 
 /**
@@ -889,7 +895,7 @@ function cptui_get_cptui_taxonomy_object( $taxonomy = '' ) {
 	if ( array_key_exists( $taxonomy, $taxonomies ) ) {
 		return $taxonomies[ $taxonomy ];
 	}
-	return '';
+	return [];
 }
 
 /**